+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "Bulk Problem Diagnostics" extension will be documented in this file.
 
-## [1.0]
+## [1.0.1]
+
+- Open only files with problems
+- Added settings to match problems severity and error message
+
+## [1.0.0]
 
 - Initial release